Update elements-min.js
Introduce a button with the "bmc-button" class that provides a seamless way for supporters to contribute. This update enriches the interface with a visually appealing and inviting feature, making it easy for users to express their support for the project. It fosters a sense of community and appreciation, turning every click into a joyful experience.